#82 FB対応 GitHubSearchAPIErrorのstruct名の修正とGitHubSearchResponse内にstructを宣言して不要なクラスを削除
