+++ iOSReferenceRepository/Model/Model/GitHubSearch/GitHubSearchAPIError.swift
@@ -9,13 +9,13 @@
 import Foundation
 
 enum GitHubSearchAPIError: Error {
-    case notModified(info: ErrorInfo) // 304
-    case forbidden(info: ErrorInfo) // 403
-    case resourceNotFound(info: ErrorInfo) // 404
-    case validationFailed(info: ErrorInfo) // 422
-    case serviceUnavailable(info: ErrorInfo) // 503
+    case notModified(info: Info) // 304
+    case forbidden(info: Info) // 403
+    case resourceNotFound(info: Info) // 404
+    case validationFailed(info: Info) // 422
+    case serviceUnavailable(info: Info) // 503
 
-    struct ErrorInfo: Codable {
+    struct Info: Codable {
         let message: String
         let errors: [Errors]?
         let documentationUrl: String
